[5.2.x] draft: разделение компонентов по файлам
