+++ CHANGELOG.md
@@ -59,7 +59,7 @@
 * 5.2.1 upd: конкретизированы типы TS
 * 5.2.1 upd: компоненты клиента переписаны на функциональные
 * 5.2.2 upd: переработанный глобальный стейт поставляется через контекст
-* 5.2.2 upd: дальнейший рефакторинг внешнего стейта и выделение презентационных компонентов
+* 5.2.2 upd: дальнейший рефакторинг внешнего стейта, выделение презентационных компонентов
 
 # Запланированные изменения:
 * **БД**: перейти с MySQL на Postgres, сохранив демонстрационный дамп с ~1000 заметок
