+++ ReducedSetSearchEngineService.sln
@@ -29,13 +29,13 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "client.jsx", "client.jsx", "{F63409DD-059D-445B-95EA-897A6333935C}"
 	ProjectSection(SolutionItems) = preProject
 		src\Rsse.Client\ClientApp\src\index.tsx = src\Rsse.Client\ClientApp\src\index.tsx
-		src\Rsse.Client\ClientApp\src\jsx\components\catalog.component.tsx = src\Rsse.Client\ClientApp\src\jsx\components\catalog.component.tsx
-		src\Rsse.Client\ClientApp\src\jsx\components\create.component.tsx = src\Rsse.Client\ClientApp\src\jsx\components\create.component.tsx
-		src\Rsse.Client\ClientApp\src\jsx\components\login.component.tsx = src\Rsse.Client\ClientApp\src\jsx\components\login.component.tsx
-		src\Rsse.Client\ClientApp\src\jsx\components\read.component.tsx = src\Rsse.Client\ClientApp\src\jsx\components\read.component.tsx
-		src\Rsse.Client\ClientApp\src\jsx\components\update.component.tsx = src\Rsse.Client\ClientApp\src\jsx\components\update.component.tsx
+		src\Rsse.Client\ClientApp\src\jsx\components\create.container.tsx = src\Rsse.Client\ClientApp\src\jsx\components\create.container.tsx
+		src\Rsse.Client\ClientApp\src\jsx\components\update.container.tsx = src\Rsse.Client\ClientApp\src\jsx\components\update.container.tsx
 		src\Rsse.Client\ClientApp\src\jsx\components\menu.component.tsx = src\Rsse.Client\ClientApp\src\jsx\components\menu.component.tsx
 		src\Rsse.Client\ClientApp\src\jsx\common\loader.tsx = src\Rsse.Client\ClientApp\src\jsx\common\loader.tsx
+		src\Rsse.Client\ClientApp\src\jsx\components\catalog.container.tsx = src\Rsse.Client\ClientApp\src\jsx\components\catalog.container.tsx
+		src\Rsse.Client\ClientApp\src\jsx\components\login.container.tsx = src\Rsse.Client\ClientApp\src\jsx\components\login.container.tsx
+		src\Rsse.Client\ClientApp\src\jsx\components\read.container.tsx = src\Rsse.Client\ClientApp\src\jsx\components\read.container.tsx
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "client.scripts", "client.scripts", "{40D75148-337E-4545-94AB-2D06AB6F27E9}"
